Made style changes to search screen. inspired by revvy carpool
