+++ frontend/ios/GetMeHome/Views/Search Section/SearchSection.swift
@@ -1,0 +1,20 @@
+//
+//  SearchSection.swift
+//  GetMeHome
+//
+//  Created by Ronald Jabouin on 1/14/24.
+//
+
+import SwiftUI
+
+struct SearchSection: View {
+    var body: some View {
+        VStack {
+            
+        }
+    }
+}
+
+#Preview {
+    SearchSection()
+}
